+++ R/glance.dic.R
@@ -15,6 +15,7 @@
 #' @export
 #' @examples
 #' library(rjags)
+#' example_jags_model$recompile()
 #' model_dic <- dic.samples(example_jags_model, n.iter = 1000)
 #' glance(model_dic)
 glance.dic <- function(x, ...){