+++ NEWS.md
@@ -3,7 +3,7 @@ mmcc 0.0.8.9400 (2019/02/01)
 
 ## New Features
 
-* `mcmc_to_dt` is now and S3 method, removing `mcmc_to_dt_stan()`.
+* `mcmc_to_dt` is now and S3 method, removing `mcmc_to_dt_stan()`, it just works on stan models now. And `mcmc_to_dt.stan()` is about 5 times faster.
 
 ## Minor Changes